pyproject.toml
[project]
name = "online-cinema-api"
version = "0.1.0"
description = "FastAPI cinema booking API"
authors = [
    {name = "Arsen Myroniuk", email = "arisenmyr@gmail.com"},
    {name = "Yehor Ripa"},
    {name = "Arsenii Aristov"},
    {name = "Bohdan Havryliyk"},
    {name = "Danylo Trukhanov"}
]
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "annotated-types (==0.7.0)",
    "anyio (==4.9.0)",
    "backports-tarfile (==1.2.0)",
    "build (==1.2.2.post1)",
    "cachecontrol (==0.14.2)",
    "certifi (==2025.1.31)",
    "charset-normalizer (==3.4.1)",
    "cleo (==2.1.0)",
    "click (==8.1.8)",
    "colorama (==0.4.6)",
    "crashtest (==0.4.1)",
    "distlib (==0.3.9)",
    "dulwich (==0.22.8)",
    "exceptiongroup (==1.2.2)",
    "fastapi (==0.115.11)",
    "fastjsonschema (==2.21.1)",
    "filelock (==3.18.0)",
    "findpython (==0.6.3)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "importlib-metadata (==8.6.1)",
    "installer (==0.7.0)",
    "jaraco-classes (==3.4.0)",
    "jaraco-context (==6.0.1)",
    "jaraco-functools (==4.1.0)",
    "keyring (==25.6.0)",
    "more-itertools (==10.6.0)",
    "msgpack (==1.1.0)",
    "packaging (==24.2)",
    "pbs-installer (==2025.3.17)",
    "pkginfo (==1.12.1.2)",
    "platformdirs (==4.3.7)",
    "poetry (==2.1.1)",
    "poetry-core (==2.1.1)",
    "pydantic (==2.10.6)",
    "pydantic-core (==2.27.2)",
    "pyproject-hooks (==1.2.0)",
    "pywin32-ctypes (==0.2.3)",
    "rapidfuzz (==3.12.2)",
    "requests (==2.32.3)",
    "requests-toolbelt (==1.0.0)",
    "shellingham (==1.5.4)",
    "sniffio (==1.3.1)",
    "starlette (==0.46.1)",
    "tomli (==2.2.1)",
    "tomlkit (==0.13.2)",
    "trove-classifiers (==2025.3.19.19)",
    "typing-extensions (==4.12.2)",
    "urllib3 (==2.3.0)",
    "uvicorn (==0.34.0)",
    "virtualenv (==20.29.3)",
    "zipp (==3.21.0)",
    "zstandard (==0.23.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.2"
black = "^25.1.0"

